Fix up MPG123.
The MPG123 maintainer kindly fixed the MPG123_EXPORT issue.
https://sourceforge.net/p/mpg123/bugs/239/

The latest release (1.23.6) also contains a def file, so all the gendef stuff
has been binned. Woop!

1.23.6 also changes `typedef long ssize_t` to `typedef ptrditt_t ssize_t`.
This slightly breaks my hack for getting MPG123 and libuv to live together,
and produces an extra warning on x64, but x64 has a lot of warnings about
size_t as well anyway, so these should be fixed properly.

Listening to MP3s still seems to work on x86 and x64!

+++ src/io.cpp
@@ -26,7 +26,7 @@
 #undef UNICODE
 // Use the same ssize_t as libmpg123 on Windows.
 #ifdef _MSC_VER
-typedef long ssize_t;
+typedef ptrdiff_t ssize_t;
 #define _SSIZE_T_
 #define _SSIZE_T_DEFINED
 #endif
